make forward.py, reformat code
